backend/requirements.txt
Flask
Flask-Cors
SQLAlchemy
nltk
spacy
transformers
numpy
pandas
psycopg2-binary
SpeechRecognition
pydub
flask-jwt-extended
flask-bcrypt
google-generativeai

# pip install --upgrade pip

# source myenv/Scripts/activate